feat: CI job build boxes to ansible runner

See merge request cf_homelab/vagrant_boxes!3

+++ build.sh
@@ -1,8 +1,23 @@
 #!/bin/bash -ex
 
-vagrant box remove "$BOX_NAME" --force || true
+if [ -z "${BOX_NAME:-}" ]; then
+    echo "BOX_NAME is required." >&2
+    exit 1
+fi
+
+template_file="$BOX_NAME.pkr.hcl"
+
+if [ ! -f "$template_file" ]; then
+    echo "Packer template '$template_file' does not exist." >&2
+    exit 1
+fi
+
+# Use a directory on the main filesystem instead of /tmp (which may be limited)
+export TMPDIR="$(pwd)/build-tmp"
+mkdir -p "$TMPDIR"
+
 curl -fsSL https://raw.githubusercontent.com/hashicorp/vagrant/main/keys/vagrant.pub -o vagrant.pub
-packer init "$BOX_NAME".pkr.hcl
+packer init "$template_file"
 PACKER_LOG=1 PACKER_LOG_PATH="packer.log" packer build "$BOX_NAME".pkr.hcl
 vagrant box add --name "$BOX_NAME" ./artifacts/"$BOX_FILE" --force
